# api/swagger/swagger.yaml
definitions:
  models.FoundVolume:
    properties:
      difference:
        description: Difference between found volume and best ask or best bid and
          found volume in percent
        type: number
      exchange:
        type: string
      index:
        description: Number of rows between found volume index and best ask or best
          bid and found volume index
        type: integer
      pair:
        type: string
      price:
        type: number
      side:
        type: string
      volume:
        type: number
      volume_time_found:
        type: string
    type: object
...
  models.Response:
    properties:
      result:
        type: string
    type: object
...
  models.UserPairs:
    properties:
      exact_value:
        example: 3
        type: number
      exchange:
        example: binance_spot
        type: string
      pair:
        example: BTC/USDT
        type: string
    type: object
info:
  contact: {}
  title: Crypto Volume Finder API
  version: "1.0"
paths:
...
  /api/user/pair:
    delete:
      consumes:
      - application/json
      description: Remove an existing pair for the authenticated user
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: The pair that should be deleted
        in: query
        name: pair
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response indicating the pair was deleted
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a user pair
      tags:
      - user-pairs
  /api/user/pair/add:
    post:
      consumes:
      - application/json
      description: Create a new pair for the authenticated user
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User pair data
        in: body
        name: pair
        required: true
        schema:
          $ref: '#/definitions/models.UserPairs'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response indicating the pair was added
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Add a new user pair
      tags:
      - user-pairs
  /api/user/pair/all-pairs:
    get:
      description: Get all user pairs associated with the authenticated user's account
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user pairs
          schema:
            items:
              $ref: '#/definitions/models.UserPairs'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Retrieve all pairs for the authenticated user
      tags:
      - user-pairs
  /api/user/pair/found-volumes:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a list of all found volumes associated
        with the authenticated user.
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/models.FoundVolume'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Retrieve all found volumes for the authenticated user
      tags:
      - user-pairs
  /api/user/pair/update-exact-value:
    put:
      consumes:
      - application/json
      description: Update an existing pair for the authenticated user
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User pair data
        in: body
        name: pair
        required: true
        schema:
          $ref: '#/definitions/models.UserPairs'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response indicating the pair was updated
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update the exact value of a user pair
      tags:
      - user-pairs
...
swagger: "2.0"